+++ App.tsx
@@ -1,24 +1,21 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
 import { Provider as PaperProvider } from 'react-native-paper';
-import TButton from './src/components/Tbutton';
+import { User, JsKeyManager } from 'tonomy-id-sdk';
+import { useEffect } from 'react';
+import MainNavigation from './src/navigation/main';
+
+
+
 export default function App() {
+
+  useEffect(() => {
+
+
+
+  })
   return (
     <PaperProvider>
-      <View style={styles.container}>
-        <Text>Open up App.tsx to start working on your app! </Text>
-        <TButton>test</TButton>
-        <StatusBar style="auto" />
-      </View>
+      <MainNavigation></MainNavigation>
     </PaperProvider>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
