+++ src/navigation/main.tsx
@@ -1,0 +1,19 @@
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import HomeScreen from '../screens/homeScreen';
+import TestScreen from '../screens/testScreen';
+
+
+const Stack = createNativeStackNavigator();
+
+export default function MainNavigation() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="home">
+        <Stack.Screen name="home" component={HomeScreen} />
+        <Stack.Screen name="test" component={TestScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  )
+}
+
