+++ src/screens/testScreen.tsx
@@ -1,0 +1,25 @@
+import TButton from '../components/Tbutton';
+import { StyleSheet, Text, View } from 'react-native';
+import { NavigationProp } from '@react-navigation/native';
+export default function HomeScreen({ navigation }: { navigation: NavigationProp<any> }) {
+
+
+
+  return (
+    <View style={styles.container}>
+
+      <Text>test Page </Text>
+      <TButton onPress={() => navigation.navigate('home')}>Go to Home</TButton>
+
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
